Implement & test registering against the user database
The way I duplicated the code necessary to isolate each tests into the
own copy of the database is very repetititve and the some refactoring
needs to happen.

Also the way the configuration is handled is leading to too many
parameters. And the configuration is currently lacking the options
necessary to connect to a different server or provide username & passwords.