Make get_auth_async a method of XHauntComponent
I needed it to be a method of an object so I could override it with
unittest.mock.patch. When it was just a function import had already
imported the original version and I couldn't change the function body.

Also moving it did let me shorten the function signature which was
also too long.